requirements.txt
pandas>=2.0
openpyxl>=3.1
sqlalchemy>=2.0
psycopg2-binary>=2.9
python-dotenv>=1.0
scikit-learn>=1.3
scipy>=1.11
numpy>=1.24
matplotlib>=3.7
seaborn>=0.12
django>=4.2
djangorestframework>=3.14
django-cors-headers>=4.0
